requirements.txt
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
croniter==0.3.30
decorator==4.4.0
Django
django-admin-rq
django-debug-toolbar
django-rq
djangorestframework
Faker
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.2
redis
requests==2.22.0
rq==1.1.0
rq-scheduler==0.9
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.2
urllib3==1.25.6
wcwidth==0.1.7
